Adding EBS Config script for Mongo instances.
This should be run to setup the RAID 10 and mount it on /data. It will (
eventually) include backup scripts and scheduling as well. Right now val
ues are hardcoded (e.g. size = 8gb and drive names = sdf, etc.), so thos
e will have to be fixed ultimately as well. I'm not sure if we want to r
un Mongo here as well, but it's certainly an option.